refactor(unused-block-params): rename to no-unused-block-params
