+++ lib/config/recommended.js
@@ -16,6 +16,6 @@ module.exports = {
     'deprecated-inline-view-helper': true,
     'deprecated-render-helper': true,
     'simple-unless': true,
-    'unused-block-params': true
+    'no-unused-block-params': true
   }
 };
