+++ lib/rules/index.js
@@ -23,7 +23,7 @@ module.exports = {
   'no-invalid-interactive': require('./lint-no-invalid-interactive'),
   'style-concatenation': require('./lint-style-concatenation'),
   'simple-unless': require('./lint-simple-unless'),
-  'unused-block-params': require('./lint-unused-block-params'),
+  'no-unused-block-params': require('./lint-no-unused-block-params'),
   'no-duplicate-attributes': require('./lint-duplicate-attributes'),
   'linebreak-style': require('./lint-linebreak-style'),
   'attribute-indentation': require('./lint-attribute-indentation'),
