+++ test/unit/rules/lint-no-unused-block-params-test.js
@@ -3,7 +3,7 @@
 const generateRuleTests = require('../../helpers/rule-test-harness');
 
 generateRuleTests({
-  name: 'unused-block-params',
+  name: 'no-unused-block-params',
 
   config: true,
